Moved both get_model to a single method inside backbone_factory, change tf.v1 Gfile ops to tf2
